Revert "simplified code and add comments"
This reverts commit de9ce2993ef9374521d9705006641e3808ed90c8.
