+++ popcnt_amd64.s
@@ -1,4 +1,4 @@
-// +build amd64,!appengine
+// +build !appengine
 
 TEXT ·hasAsm(SB),4,$0-1
 MOVQ $1, AX
